refactor(employee): streamline EditEmployeePage and extract EditEmployeeForm component for better organization (review fix)
